+++ src/Nethermind/Nethermind.Evm/Instructions/EvmInstructions.ControlFlow.cs
@@ -4,6 +4,9 @@
 using System;
 using System.Runtime.CompilerServices;
 using System.Runtime.Intrinsics;
+using Nethermind.Core.Specs;
+using Nethermind.Core;
+using Nethermind.State;
 
 namespace Nethermind.Evm;
 using Int256;
@@ -120,6 +123,157 @@ internal sealed partial class EvmInstructions
     InvalidJumpDestination:
         return EvmExceptionType.InvalidJumpDestination;
     }
+
+    /// <summary>
+    /// Stops the execution of the EVM.
+    /// In EOFCREATE or TXCREATE executions, the STOP opcode is considered illegal.
+    /// </summary>
+    [SkipLocalsInit]
+    public static EvmExceptionType InstructionStop(VirtualMachine vm, ref EvmStack stack, ref long gasAvailable, ref int programCounter)
+    {
+        // In contract creation contexts, a STOP is not permitted.
+        if (vm.EvmState.ExecutionType is ExecutionType.EOFCREATE or ExecutionType.TXCREATE)
+        {
+            return EvmExceptionType.BadInstruction;
+        }
+
+        return EvmExceptionType.Stop;
+    }
+
+    /// <summary>
+    /// Implements the REVERT opcode.
+    /// Pops a memory offset and length from the stack, updates memory gas cost, loads the return data,
+    /// and returns a revert exception.
+    /// </summary>
+    [SkipLocalsInit]
+    public static EvmExceptionType InstructionRevert(VirtualMachine vm, ref EvmStack stack, ref long gasAvailable, ref int programCounter)
+    {
+        // Attempt to pop memory offset and length; if either fails, signal a stack underflow.
+        if (!stack.PopUInt256(out UInt256 position) ||
+            !stack.PopUInt256(out UInt256 length))
+        {
+            goto StackUnderflow;
+        }
+
+        // Ensure sufficient gas for any required memory expansion.
+        if (!UpdateMemoryCost(vm.EvmState, ref gasAvailable, in position, in length))
+        {
+            goto OutOfGas;
+        }
+
+        // Copy the specified memory region as return data.
+        vm.ReturnData = vm.EvmState.Memory.Load(in position, in length).ToArray();
+
+        return EvmExceptionType.Revert;
+    // Jump forward to be unpredicted by the branch predictor.
+    OutOfGas:
+        return EvmExceptionType.OutOfGas;
+    StackUnderflow:
+        return EvmExceptionType.StackUnderflow;
+    }
+
+    /// <summary>
+    /// Executes the SELFDESTRUCT opcode.
+    /// This method handles gas adjustments, account balance transfers,
+    /// and marks the executing account for destruction.
+    /// </summary>
+    [SkipLocalsInit]
+    private static EvmExceptionType InstructionSelfDestruct(VirtualMachine vm, ref EvmStack stack, ref long gasAvailable, ref int programCounter)
+    {
+        // Increment metrics for self-destruct operations.
+        Metrics.IncrementSelfDestructs();
+
+        EvmState vmState = vm.EvmState;
+        IReleaseSpec spec = vm.Spec;
+        IWorldState state = vm.WorldState;
+
+        // SELFDESTRUCT is forbidden during static calls.
+        if (vmState.IsStatic)
+            goto StaticCallViolation;
+
+        // If Shanghai DDoS protection is active, charge the appropriate gas cost.
+        if (spec.UseShanghaiDDosProtection)
+        {
+            gasAvailable -= GasCostOf.SelfDestructEip150;
+        }
+
+        // Pop the inheritor address from the stack; signal underflow if missing.
+        Address inheritor = stack.PopAddress();
+        if (inheritor is null)
+            goto StackUnderflow;
+
+        // Charge gas for account access; if insufficient, signal out-of-gas.
+        if (!ChargeAccountAccessGas(ref gasAvailable, vm, inheritor, chargeForWarm: false))
+            goto OutOfGas;
+
+        Address executingAccount = vmState.Env.ExecutingAccount;
+        bool createInSameTx = vmState.AccessTracker.CreateList.Contains(executingAccount);
+        // Mark the executing account for destruction if allowed.
+        if (!spec.SelfdestructOnlyOnSameTransaction || createInSameTx)
+            vmState.AccessTracker.ToBeDestroyed(executingAccount);
+
+        // Retrieve the current balance for transfer.
+        UInt256 result = state.GetBalance(executingAccount);
+        if (vm.TxTracer.IsTracingActions)
+            vm.TxTracer.ReportSelfDestruct(executingAccount, result, inheritor);
+
+        // For certain specs, charge gas if transferring to a dead account.
+        if (spec.ClearEmptyAccountWhenTouched && !result.IsZero && state.IsDeadAccount(inheritor))
+        {
+            if (!UpdateGas(GasCostOf.NewAccount, ref gasAvailable))
+                goto OutOfGas;
+        }
+
+        // If account creation rules apply, ensure gas is charged for new accounts.
+        bool inheritorAccountExists = state.AccountExists(inheritor);
+        if (!spec.ClearEmptyAccountWhenTouched && !inheritorAccountExists && spec.UseShanghaiDDosProtection)
+        {
+            if (!UpdateGas(GasCostOf.NewAccount, ref gasAvailable))
+                goto OutOfGas;
+        }
+
+        // Create or update the inheritor account with the transferred balance.
+        if (!inheritorAccountExists)
+        {
+            state.CreateAccount(inheritor, result);
+        }
+        else if (!inheritor.Equals(executingAccount))
+        {
+            state.AddToBalance(inheritor, result, spec);
+        }
+
+        // Special handling when SELFDESTRUCT is limited to the same transaction.
+        if (spec.SelfdestructOnlyOnSameTransaction && !createInSameTx && inheritor.Equals(executingAccount))
+            goto Stop; // Avoid burning ETH if contract is not destroyed per EIP clarification
+
+        // Subtract the balance from the executing account.
+        state.SubtractFromBalance(executingAccount, result, spec);
+
+    // Jump forward to be unpredicted by the branch predictor.
+    Stop:
+        return EvmExceptionType.Stop;
+    OutOfGas:
+        return EvmExceptionType.OutOfGas;
+    StackUnderflow:
+        return EvmExceptionType.StackUnderflow;
+    StaticCallViolation:
+        return EvmExceptionType.StaticCallViolation;
+    }
+
+    /// <summary>
+    /// Handles invalid opcodes by deducting a high gas cost and returning a BadInstruction error.
+    /// </summary>
+    public static EvmExceptionType InstructionInvalid(VirtualMachine _, ref EvmStack stack, ref long gasAvailable, ref int programCounter)
+    {
+        gasAvailable -= GasCostOf.High;
+        return EvmExceptionType.BadInstruction;
+    }
+
+    /// <summary>
+    /// Default handler for undefined opcodes, always returning a BadInstruction error.
+    /// </summary>
+    public static EvmExceptionType InstructionBadInstruction(VirtualMachine _, ref EvmStack stack, ref long gasAvailable, ref int programCounter)
+        => EvmExceptionType.BadInstruction;
 
     /// <summary>
     /// Validates a jump destination and, if valid, updates the program counter.
